Moved sensor stuff to separate module
